// core/ec-core/umbracoClient/models/ContentPanePropertiesModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ApiBlockListModel } from './ApiBlockListModel';
import type { ApiLinkModel } from './ApiLinkModel';
export type ContentPanePropertiesModel = {
    headline?: string | null;
    link?: Array<ApiLinkModel> | null;
    largeImage?: ApiBlockListModel;
    topSmallImage?: ApiBlockListModel;
    bottomSmallImage?: ApiBlockListModel;
};