// core/ec-core/umbracoClient/models/ImageWithLinkPropertiesModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ApiLinkModel } from './ApiLinkModel';
import type { IApiMediaWithCropsModel } from './IApiMediaWithCropsModel';
export type ImageWithLinkPropertiesModel = {
    image?: Array<IApiMediaWithCropsModel> | null;
    link?: Array<ApiLinkModel> | null;
};